make driver close more gracefully on device hard remove
Check removed flag in polling loops, indirect rr/wr
and in mcu_send() to make sure logs are not spammed
with failure messages.

Also set the removed flag during failed Tx submit.
If the device is mostly idle this is likely to be
the first point we will notice it disappeared
because mac80211 will try to send disassoc frame.

Signed-off-by: Jakub Kicinski <kubakici@wp.pl>